requirements.txt
flask
werkzeug
deepface
mtcnn
pillow
numpy
